new_dashboard/requirements.txt
flask
flask-socketio
flask-wtf
flask-talisman
flask-limiter
psycopg2-binary
requests
python-dotenv
itsdangerous
requests==2.31.0
gunicorn==21.2.0
Werkzeug==3.0.1
